% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fold_similarity.r
\name{cluster_fold_similarity}
\alias{cluster_fold_similarity}
\title{Calculate cluster similarity between clusters from different single cell samples.}
\usage{
cluster_fold_similarity(
  sce_list = NULL,
  sample_names = NULL,
  top_n = 1,
  top_n_genes = 1
)
}
\arguments{
\item{sce_list}{List. A list of single cell experiments. At least 2 single cell experiments are needed. The objects are expected to be subsets of the original data, 
containing the selected features (e.g. 2000 most variable genes).}

\item{sample_names}{Character Vector. Specify the sample names, if not a number corresponding with its position on (sce_list).}

\item{top_n}{Numeric. Specifies the number of target clusters with best similarity to report for each cluster comparison (default 1).}

\item{top_n_genes}{Numeric. Number of top genes that explains the clusters similarity to report for each cluster comparison (default 1).}
}
\value{
The function returns a \link[S4Vectors]{DataFrame} containing the best top similarities between all possible pairs of single cell samples.
similarity_value=integer(): The top similarity value calculated between sample_l:cluster_l and sample_r. 
sem=integer(): Standar Error of the Mean (SEM) of the mean of the values of the coeficient calculated for all genes.
sample_l=integer(): Sample left, the sample source which is being compared. 
cluster_l=integer(): Cluster left, the cluster source which is being compared.
sample_r=integer(): Sample right, the sample target which is being compared with sample_l.
cluster_r=integer(): Cluster right, the cluster target which is being compared with cluster_l.
top_gene_conserved=character(): The gene that showed most similar between cluster_l & cluster_r.
}
\description{
`cluster_fold_similarity()` returns a dataframe containing the best top similarities between all possible pairs of single cell samples.
}
\details{
This function will calculate a similarity coeficient using the fold changes of shared genes among clusters of different samples/batches. The similarity coeficient
is calculated using the dotproduct of every pairwise combination of Fold Changes between a source cluster i of sample n and all the target clusters in sample j.
}
