% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldchange_composition.r
\name{foldchangeComposition}
\alias{foldchangeComposition}
\title{Calculate the dot product between all the possible combinations of foldchanges from diferent clusters.}
\usage{
foldchangeComposition(root = NULL, comparative = NULL)
}
\arguments{
\item{root}{Dataframe. Foldchanges between a source cluster and all the other clusters found on a sample.}

\item{comparative}{Dataframe. Foldchanges between a cluster and all the other clusters found on a second sample to be compared with the (root) cluster foldchanges.}
}
\value{
A dataframe containing the hadamard product of all the possible combinations of foldchanges.
}
\description{
`foldchangeComposition()` returns a dataframe containing the best top similarities between all possible pairs of single cell samples.
}
\details{
This function will perform the dot product of each possible combination of foldchanges, by constructing two dataframes: 
one with the source cluster's foldchanges and the other with the foldchange values of a target sample's cluster.
The computation of all the possible combinations is the hadamard product of the matrix.
}
\keyword{internal}
