% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_cluster_plot.r
\name{multi_cluster_plot}
\alias{multi_cluster_plot}
\title{Plot several single-cell datasets/samples dimensionality reductions (PCA, UMAP, tSNE) in a single plot 3D plot, each layer corresponding to a dataset/sample.}
\usage{
multi_cluster_plot(
  sce_list = NULL,
  similarity_table = NULL,
  dim_method = "PCA",
  pdf_name = NULL
)
}
\arguments{
\item{sce_list}{List. A list of single cell experiments. At least 2 single cell experiments are needed.}

\item{similarity_table}{DataFrame. Optional. A similarity table previously calculated using ClusterFoldSimilarity function.}

\item{dim_method}{Character. Specify the dimensionality reduction used, either PCA, UMAP or tSNE. Default: PCA.}

\item{pdf_name}{Character. If specify the plot is saved on a PDF file with the given name. Default: plot in the current graphic device.}
}
\value{
A dimensionality reduction 3D plot (PCA, UMAP, tSNE) of a batch of samples/datasets.
}
\description{
`multi_cluster_plot()` plot dimensionality reduction (PCA, UMAP, tSNE) of a batch of samples/datasets.
}
\details{
This function will plot a list of single-cell datasets/samples dimensionality reduction in a 3D layered plot.
}
