% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_cluster_fold_change.r
\name{pairwise_cluster_fold_change}
\alias{pairwise_cluster_fold_change}
\title{Calculate the gene mean expression Fold Change between all possible combinations of clusters.}
\usage{
pairwise_cluster_fold_change(x, clusters)
}
\arguments{
\item{x}{Dataframe. Normalized counts containint gene expression.}

\item{comparative}{Factor. A vector of corresponding cluster for each sample of (x).}
}
\value{
A list of dataframes containing the pairwise fold changes between all combinations of cluster.
}
\description{
`pairwise_cluster_fold_change()` returns a list of dataframes containing the pairwise fold changes between all combinations of cluster.
}
\details{
This function will perform fold change ratios of mean gene expression between all possible combination of clusters specified on colLabels inside the sc object.
}
\author{
Oscar Gonzalez-Velasco
}
\keyword{internal}
