% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_heatmap.r
\name{similarityHeatmap}
\alias{similarityHeatmap}
\title{Plot a heatmap of the similarity values obtained using cluster fold similarity}
\usage{
similarityHeatmap(
  similarityTable = NULL,
  mainDataset = NULL,
  otherDatasets = NULL
)
}
\arguments{
\item{similarityTable}{A DataFrame containing the similarities between all possible pairs of single cell samples obtained with \link[ClusterFoldSimilarity]{clusterFoldSimilarity} using the option n_top=Inf.}

\item{mainDataset}{Numeric. Specify the main dataset (y axis). It corresponds with the datasetL column from the similarityTable}

\item{otherDatasets}{Numeric. Specify some specific dataset to be ploted along the mainDataset (x axis, default: all other datasets found on datasetR column from similarity_table).}
}
\value{
The function returns a heatmap ggplot object.
}
\description{
`similarityHeatmap()` returns a ggplot heatmap representing the similarity values between pairs of clusters as obtained from \link[ClusterFoldSimilarity]{clusterFoldSimilarity}.
}
\details{
This function plots a heatmap using ggplot. It is intended to be used with the output table from \link[ClusterFoldSimilarity]{clusterFoldSimilarity}, which includes the columns: datasetL (the dataset used for comparison)
datasetR (the dataset against datasetL has been contrasted), clusterL (clusters from datasetL), clusterR (clusters from datasetR) and the similarityValue.
}
\examples{
if (requireNamespace("Seurat") & requireNamespace("SingleCellExperiment") & requireNamespace("scRNAseq") & requireNamespace("SeuratObject")){
library(ClusterFoldSimilarity)
# Mouse brain single-cell RNA-seq 1 from Romanov et. al.
mouse.brain.romanov <- scRNAseq::RomanovBrainData(ensembl = TRUE)
colnames(mouse.brain.romanov) <- colData(mouse.brain.romanov)$cellID
rownames(colData(mouse.brain.romanov)) <- colData(mouse.brain.romanov)$cellID
singlecell.1.seurat <- CreateSeuratObject(counts = counts(mouse.brain.romanov),meta.data = as.data.frame(colData(mouse.brain.romanov)))

# Mouse brain single-cell RNA-seq 2 from Zeisel et. al.
mouse.brain.zei <- scRNAseq::ZeiselBrainData(ensembl = TRUE)
singlecell.2.seurat <- CreateSeuratObject(counts = counts(mouse.brain.zei),meta.data = as.data.frame(colData(mouse.brain.zei)))

# Create a list with the unprocessed single-cell datasets
singlecell.object.list <- list(singlecell.1.seurat,singlecell.2.seurat)
# Apply the same processing to each dataset and return a list of single-cell analysis
singlecell.object.list <- lapply(X = singlecell.object.list, FUN = function(x){
  x <- NormalizeData(x)
  x <- FindVariableFeatures(x, selection.method = "vst", nfeatures = 1000)
  x <- ScaleData(x,features = VariableFeatures(x))
  x <- RunPCA(x, features = VariableFeatures(object = x))
  x <- FindNeighbors(x, dims = seq(10))
  x <- FindClusters(x, resolution = 0.1)
})
# singlecell.object.list consist of 2 or more single-cell objects
# Using topN = Inf by default plots a heatmap using the similarity values: 
similarity.table.all <- clusterFoldSimilarity(sceList=singlecell.object.list, topN=Inf)
# Using the dataset 2 as a reference on the Y-axis of the heatmap:
similarityHeatmap(similarityTable=similarity.table.all, mainDataset=2)
}

}
\author{
Oscar Gonzalez-Velasco
}
