% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fold_similarity.r
\name{cluster_fold_similarity}
\alias{cluster_fold_similarity}
\title{Calculate cluster similarity between clusters from different single cell samples.}
\usage{
cluster_fold_similarity(
  sce_list = NULL,
  sample_names = NULL,
  top_n = 1,
  top_n_genes = 1
)
}
\arguments{
\item{sce_list}{List. A list of single cell experiments. At least 2 single cell experiments are needed. The objects are expected to be subsets of the original data, 
containing the selected features (e.g. 2000 most variable genes).}

\item{sample_names}{Character Vector. Specify the sample names, if not a number corresponding with its position on (sce_list).}

\item{top_n}{Numeric. Specifies the number of target clusters with best similarity to report for each cluster comparison (default 1). If set to Inf, then all similarity values from all possible pairs of clusters are returned.}

\item{top_n_genes}{Numeric. Number of top genes that explains the clusters similarity to report for each cluster comparison (default 1). If top_n = Inf then top_n_genes is automatically set to 1.}
}
\value{
The function returns a \linkS4class{DataFrame} containing the best top similarities between all possible pairs of single cell samples. Column values are:
\tabular{ll}{
   \code{similarity_value} \tab The top similarity value calculated between dataset_l:cluster_l and dataset_r. \cr
   \tab \cr
   \code{sem} \tab Standar Error of the Mean (SEM) of the mean of the values of the coeficient calculated for all genes. \cr
   \tab \cr
   \code{dataset_l} \tab Dataset left, the dataset/sample which has been used to be compared.  \cr
   \tab \cr
   \code{cluster_l} \tab Cluster left, the cluster source from dataset_l which has been compared. \cr
   \tab \cr
   \code{dataset_r} \tab Dataset right, the dataset/sample used for comparison against dataset_l. \cr
   \tab \cr
   \code{cluster_r} \tab Cluster right, the cluster target from dataset_r which is being compared with the cluster_l from dataset_l. \cr
   \tab \cr
   \code{top_gene_conserved} \tab The gene that showed most similar between cluster_l & cluster_r. \cr
}
}
\description{
`cluster_fold_similarity()` returns a dataframe containing the best top similarities between all possible pairs of single cell samples.
}
\details{
This function will calculate a similarity coeficient using the fold changes of shared genes among clusters of different samples/batches. The similarity coeficient
is calculated using the dotproduct of every pairwise combination of Fold Changes between a source cluster i of sample n and all the target clusters in sample j.
}
\examples{
library(Seurat)
library(scRNAseq)
library(ClusterFoldSimilarity)
# Mouse brain single-cell RNA-seq 1 from Romanov et. al.
mouse.brain.romanov <- scRNAseq::RomanovBrainData(ensembl = TRUE)
colnames(mouse.brain.romanov) <- colData(mouse.brain.romanov)$cellID
rownames(colData(mouse.brain.romanov)) <- colData(mouse.brain.romanov)$cellID
singlecell.1.seurat <- CreateSeuratObject(counts = counts(mouse.brain.romanov),meta.data = as.data.frame(colData(mouse.brain.romanov)))

# Mouse brain single-cell RNA-seq 2 from Zeisel et. al.
mouse.brain.zei <- scRNAseq::ZeiselBrainData(ensembl = TRUE)
singlecell.2.seurat <- CreateSeuratObject(counts = counts(mouse.brain.zei),meta.data = as.data.frame(colData(mouse.brain.zei)))

# Create a list with the unprocessed single-cell datasets
singlecell.object.list <- list(singlecell.1.seurat,singlecell.2.seurat)
# Apply the same processing to each dataset and return a list of single-cell analysis
singlecell.object.list <- lapply(X = singlecell.object.list, FUN = function(x){
  x <- NormalizeData(x)
  x <- FindVariableFeatures(x, selection.method = "vst", nfeatures = 1000)
  x <- ScaleData(x,features = VariableFeatures(x))
  x <- RunPCA(x, features = VariableFeatures(object = x))
  x <- FindNeighbors(x, dims = seq(10))
  x <- FindClusters(x, resolution = 0.1)
})
# Calculate the similarity between the two datasets:
similarity.table <- cluster_fold_similarity(sce_list = singlecell.object.list,top_n = 1)
head(similarity.table)

}
\author{
Oscar Gonzalez-Velasco
}
