% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters_graph.r
\name{plot_clusters_graph}
\alias{plot_clusters_graph}
\title{Creates a graph plot using the similarity values calculated with ClusterFoldSimilarity().}
\usage{
plot_clusters_graph(similarity.table = NULL)
}
\arguments{
\item{similarity.table}{Dataframe. A table obtained from ClusterFoldSimilarity that contains the similarity values as a column "similarity_value" that represents 
the similarity of a source cluster to a target cluster.}
}
\value{
This function plots a graph in which the nodes are clusters from a specific dataset, the edges represent the similarity and the direction
of that similarity between clusters.
}
\description{
`plot_clusters_graph()` Creates a graph plot using the similarity values calculated with ClusterFoldSimilarity().
}
\details{
This function will calculate a similarity coeficient using the fold changes of shared genes among clusters of different samples/batches. The similarity coeficient
is calculated using the dotproduct of every pairwise combination of Fold Changes between a source cluster i of sample n and all the target clusters in sample j.
}
