% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_heatmap.r
\name{similarity_heatmap}
\alias{similarity_heatmap}
\title{Plot a heatmap of the similarity values obtained using cluster fold similarity}
\usage{
similarity_heatmap(
  similarity_table = NULL,
  main_dataset = NULL,
  other_datasets = NULL
)
}
\arguments{
\item{similarity_table}{A DataFrame containing the similarities between all possible pairs of single cell samples obtained with \link[ClusterFoldSimilarity]{cluster_fold_similarity} using the option n_top=Inf.}

\item{main_dataset}{Numeric. Specify the main dataset (y axis). It corresponds with the dataset_l column from the similarity_table.}

\item{other_datasets}{Numeric. Specify some specific dataset to be ploted along the main_dataset (x axis, default: all other datasets found on dataset_r column from similarity_table).}
}
\value{
The function returns a heatmap ggplot object.
}
\description{
`similarity_heatmap()` returns a ggplot heatmap representing the similarity values between pairs of clusters as obtained from \link[ClusterFoldSimilarity]{cluster_fold_similarity}.
}
\details{
This function plots a heatmap using ggplot. It is intended to be used with the output table from \link[ClusterFoldSimilarity]{cluster_fold_similarity}, which includes the columns: dataset_l (the dataset used for comparison)
dataset_r (the dataset against dataset_l has been contrasted), cluster_l (clusters from dataset_l), cluster_r (clusters from dataset_r) and the similarity_value.
}
\examples{
library(Seurat)
library(scRNAseq)
library(ClusterFoldSimilarity)
# Mouse brain single-cell RNA-seq 1 from Romanov et. al.
mouse.brain.romanov <- scRNAseq::RomanovBrainData(ensembl = TRUE)
colnames(mouse.brain.romanov) <- colData(mouse.brain.romanov)$cellID
rownames(colData(mouse.brain.romanov)) <- colData(mouse.brain.romanov)$cellID
singlecell.1.seurat <- CreateSeuratObject(counts = counts(mouse.brain.romanov),meta.data = as.data.frame(colData(mouse.brain.romanov)))

# Mouse brain single-cell RNA-seq 2 from Zeisel et. al.
mouse.brain.zei <- scRNAseq::ZeiselBrainData(ensembl = TRUE)
singlecell.2.seurat <- CreateSeuratObject(counts = counts(mouse.brain.zei),meta.data = as.data.frame(colData(mouse.brain.zei)))

# Create a list with the unprocessed single-cell datasets
singlecell.object.list <- list(singlecell.1.seurat,singlecell.2.seurat)
# Apply the same processing to each dataset and return a list of single-cell analysis
singlecell.object.list <- lapply(X = singlecell.object.list, FUN = function(x){
  x <- NormalizeData(x)
  x <- FindVariableFeatures(x, selection.method = "vst", nfeatures = 1000)
  x <- ScaleData(x,features = VariableFeatures(x))
  x <- RunPCA(x, features = VariableFeatures(object = x))
  x <- FindNeighbors(x, dims = seq(10))
  x <- FindClusters(x, resolution = 0.1)
})
# singlecell.object.list consist of 2 or more single-cell objects
# Using top_n = Inf by default plots a heatmap using the similarity values: 
similarity.table.all <- cluster_fold_similarity(sce_list = singlecell.object.list,top_n = Inf)
# Ploting the dataset 2 on the Y-axis:
similarity_heatmap(similarity_table = similarity.table.all,main_dataset = 2)

}
\author{
Oscar Gonzalez-Velasco
}
